+++ CHANGELOG.md
@@ -6,8 +6,8 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ### New feature
 
-- `deleteSolidDataset`: Add a function in the `resource` module to delete a dataset, 
-mimicing the interface of `@inrupt/solid-client`'s `deleteSolidDataset`.
+- `deleteSolidDataset` and `deleteFile`: Add functions to the `resource` module
+  to delete resources, following the interface of `@inrupt/solid-client`.
 
 ## [2.3.2](https://github.com/inrupt/solid-client-access-grants-js/releases/tag/v2.3.2) - 2023-06-05